requirements.txt
Django==5.2.11
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.7
drf-yasg